Substantial fixes and improvements
In writing unit tests in node.js, I realized just how much was missing from code that was supposed to work, as well as substantial improvements in JSON returns and what not. Basically, making sure this server acted identical to the Java server, and supported a REST client just as well. Now it does. There are way more unit tests to write, and probably some more fixes, but at this point, happy paths should work 100%, and the rest at least 90%.