Broke some tests
But improved the code in essence by undoing changes to an auto-generated class. I still need to figure out how to handle the situation where there are required fields in the database that don't get passed (due to bad programming) from the client to the server. The JSON parser/POCO framework just assigns default values, and the fact that they are marked "non-nullable" in the database means I can't make them nullable in the POCO